// src/enums/damage-type.ts
export enum DamageType {
	None = 'None',
	Any = '[choose one damage type]',
	All = 'All damage types',
	Acid = 'Acid',
	Edged = 'Edged',
	Impact = 'Impact',
	Piercing = 'Piercing',
	Cold = 'Cold',
	Electricity = 'Electricity',
	Fire = 'Fire',
	Light = 'Light',
	Sonic = 'Sonic',
	Decay = 'Decay',
	Poison = 'Poison',
	Psychic = 'Psychic'
}